+++ functions.php
@@ -12,7 +12,7 @@
  * @copyright  2015 WebMan - Oliver Juhas
  *
  * @since    1.0
- * @version  1.3
+ * @version  1.4
  *
  * @link  http://www.webmandesign.eu
  *
@@ -33,23 +33,18 @@
 		$theme_data = wp_get_theme();
 
 	//Basic constants
-		if ( ! defined( 'WM_THEME_NAME' ) )            define( 'WM_THEME_NAME',            $theme_data->get( 'Name' )                                   );
-		if ( ! defined( 'WM_THEME_SHORTNAME' ) )       define( 'WM_THEME_SHORTNAME',       str_replace( array( '-lite', '-plus' ), '', get_template() ) );
-		if ( ! defined( 'WM_THEME_VERSION' ) )         define( 'WM_THEME_VERSION',         $theme_data->get( 'Version' )                                );
+		if ( ! defined( 'WM_THEME_NAME' ) )       define( 'WM_THEME_NAME',       $theme_data->get( 'Name' )                                   );
+		if ( ! defined( 'WM_THEME_SHORTNAME' ) )  define( 'WM_THEME_SHORTNAME',  str_replace( array( '-lite', '-plus' ), '', get_template() ) );
+		if ( ! defined( 'WM_THEME_VERSION' ) )    define( 'WM_THEME_VERSION',    $theme_data->get( 'Version' )                                );
 
-		if ( ! defined( 'WM_SCRIPTS_VERSION' ) )       define( 'WM_SCRIPTS_VERSION',       esc_attr( trim( WM_THEME_VERSION ) )                         );
-		if ( ! defined( 'WM_WP_COMPATIBILITY' ) )      define( 'WM_WP_COMPATIBILITY',      4.1                                                          );
+		if ( ! defined( 'WM_SCRIPTS_VERSION' ) )  define( 'WM_SCRIPTS_VERSION',  esc_attr( trim( WM_THEME_VERSION ) )                         );
+		if ( ! defined( 'WM_WP_COMPATIBILITY' ) ) define( 'WM_WP_COMPATIBILITY', 4.1                                                          );
 
 	//Dir constants
-		if ( ! defined( 'WM_INC_DIR' ) )               define( 'WM_INC_DIR',               trailingslashit( 'inc/' )                                    );
+		if ( ! defined( 'WM_INC_DIR' ) )          define( 'WM_INC_DIR',          trailingslashit( 'inc/' )                                    );
 
 	//URL constants
-		if ( ! defined( 'WM_THEME_AUTHOR_URI' ) )      define( 'WM_THEME_AUTHOR_URI',      esc_url( $theme_data->get( 'AuthorURI' ) )                   );
-
-	//Theme design constants
-		if ( ! defined( 'WM_IMAGE_SIZE_ITEMS' ) )      define( 'WM_IMAGE_SIZE_ITEMS',      'thumbnail'                                                  );
-		if ( ! defined( 'WM_IMAGE_SIZE_ITEMS_MENU' ) ) define( 'WM_IMAGE_SIZE_ITEMS_MENU', 'banner-small'                                               );
-		if ( ! defined( 'WM_IMAGE_SIZE_SINGULAR' ) )   define( 'WM_IMAGE_SIZE_SINGULAR',   'large'                                                      );
+		if ( ! defined( 'WM_THEME_AUTHOR_URI' ) ) define( 'WM_THEME_AUTHOR_URI', esc_url( $theme_data->get( 'AuthorURI' ) )                   );